Suggestion for fix

+++ GV_Assignment_2/GV_Assignment_2/Sprites.cpp
@@ -105,7 +105,7 @@ void Sprites::Render(int xPos, int yPos, double angle, Screen& screen)
         targetRect.w = m_spriteDimension.x;
         targetRect.h = m_spriteDimension.y;
 
-        SDL_Point centre{ m_spriteDimension.x * 0.5f, m_spriteDimension.y * 0.5f };
+        SDL_Point centre{ static_cast<int>(m_spriteDimension.x * 0.5f), m_spriteDimension.y * 0.5f };
 
         SDL_RenderCopyEx(screen.GetRenderer(), m_image, &sourceRect, &targetRect, angle, &centre, SDL_FLIP_NONE);
     }
